refactor(backup, pdc) (#3357)
# Description
Refactoring of backup and pdc related code.

# Added
# Changed
Renamed BackupAPI to IlluminaBackupService
Renamed PdcAPI to PdcService
Moved flow cell specific backup functions from the PdcService to the IlluminaBackupService
Moved tests
Moved fixtures
