+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 61.2.5
+current_version = 61.2.6
 commit = True
 tag = True
 tag_name = v{new_version}
