+++ cg/__init__.py
@@ -1,2 +1,2 @@
 __title__ = "cg"
-__version__ = "61.2.5"
+__version__ = "61.2.6"
