+++ tests/cli/backup/test_backup_command.py
@@ -294,7 +294,10 @@ def test_run_fetch_flow_cell_dry_run_retrieval_time(
 
     # GIVEN that the backup api returns a retrieval time
     expected_time = 60
-    mocker.patch("cg.meta.backup.backup.BackupAPI.fetch_flow_cell", return_value=expected_time)
+    mocker.patch(
+        "cg.services.illumina_services.backup_services.backup_service.IlluminaBackupService.fetch_flow_cell",
+        return_value=expected_time,
+    )
 
     # WHEN running the fetch flow cell command without specifying any flow cell in dry run mode
     result = cli_runner.invoke(fetch_flow_cell, ["--dry-run"], obj=backup_context)
